Create a sun7i-a20-bananapi.dtb hard link to bananapi.dtb to fix a boot failure on the Banana Pi SoC.
This is a direct commit to stable/11, as the sun7i-a20-bananapi.dtb
file exists in head, but appears to have been part of a larger
rework of dtb-related files that may have larger consequences than
hard link creation.  Note: creating a hard link to dtb files was
an original fix in 12-CURRENT beforehand, introduced in r319603.

Approved by:	re (marius)
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/11@333506 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f