aw_spi: Fix some silly clock mistake
The module uses the mod clock and not the ahb one.
We need to set the mod clock to twice the speed requested as the smallest
divider in the controller is 2.
The clock test function weren't calculating the register value best on the
best div but on the max one.
The cdr2 test function was using the cdr1 formula.

Pointy Hat: manu


git-svn-id: https://svn.freebsd.org/base/head@333716 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f