MFC r333737:
release: arm: Format FAT partition as FAT16

r332674 raised the size of the FAT partition from 2MB to 41MB for some
boards. But we format them in FAT12 and this size appears to be to big
for FAT12 and some SoC bootrom cannot cope with that.
Format the msdosfs partition as FAT16,

PR:             228285
Approved by:	re (marius)


git-svn-id: https://svn.freebsd.org/base/stable/11@333739 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f