MFC r333176:
Fix in the documentation that the default hop limit is not 30, but
the value of the sysctl variable net.inet6.ip6.hlim.
This is true since
https://svnweb.freebsd.org/base?view=revision&revision=122574

The default of 30 (which was correct up to r122574) was incorrectly
documented in
https://svnweb.freebsd.org/base?view=revision&revision=130268

Thanks to Timo Voelker for makeing me aware of the inconsistency
between to code and the documentation.

Approved by:	re@ (marius)


git-svn-id: https://svn.freebsd.org/base/stable/11@334442 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f