top(1): Display of TID when using 'H' flag
Some users prefer seeing the TID when viewing individual threads. This
makes sense as the PID will be the same for multiple entries. An attempt
was made to include both, but there is insufficient room. As such, using
the TID.

While here, rename the header variables to be more understandable.

Discussed with:	mmacy
Reported on:	2009-10-07


git-svn-id: https://svn.freebsd.org/base/head@334474 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f