indent(1): restore working -pcs
My previous indent(1) commit accidentally broke the -pcs option (which adds
space between function name and opening parenthesis in function calls) by
copying all but one of a few conditions in an if clause. Reinstate the
condition.

Add a regression test to lower the chances of breaking it again.

Correct a comment with description of what the option does.


git-svn-id: https://svn.freebsd.org/base/head@334493 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f