Finish COMPAT_AOUT support for amd64.  It wasn't in any amd64 or MI file in /sys/conf, so was unavailable in configurations that don't use modules, and was not testable or notable in NOTES.  Its normal configuration (not using a module) is still silently deprecated in aout(4) by not mentioning it there.
Update i386 NOTES for COMPAT_AOUT.  It is not i386-only, or even very MD.
Sort its entry better.

Finish gzip configuration (but not support) for amd64.  gzip is really
gzipped aout.  It is currently broken even for i386 (a call to vm fails).
amd64 has always attempted to configure and test it, but it depends on
COMPAT_AOUT (as noted).  The bug that it depends on unconfigured files
was not detected since it is configured as a device.  All other optional
image activators are configured properly using an option.


git-svn-id: https://svn.freebsd.org/base/head@334523 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f