Improve defaults for per-CPU kernel console colors, especially with 2 or 4 CPUs.  Add a compile-time option SC_KERNEL_CONS_ATTRS to control the defaults.
Default to color numbers in reverse order to CPU numbers (instead of
in the same order with white first and wrapping to dark grey), so that
the brightest bright colors are used first.  Don't use dark grey at all;
replace it by dark green.

Syscons has too many compile-time options, but this one is needed in
in case the defaults give something like white on white, or the user
really hates this feature and can't wait to turn it off in rc.

MFC after:	next release?


git-svn-id: https://svn.freebsd.org/base/head@334530 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f