Fix a gcc8 warning about a write only variable.
gcc8 warns that "verf" was set but not used. This was because the code
that uses it is disabled via a "#if 0".
This patch adds a "#if 0" to the variable's declaration and assignment
to get rid of the warning.
This way the code could be re-enabled without difficulty.

Requested by:	mmacy
MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@334586 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f