MFC r333603:
Ensure that the MTU's used are multiple of 4.

The length of SCTP packets is always a multiple of 4. Therefore,
ensure that the MTUs used are also a multiple of 4.

Thanks to Irene Ruengeler for providing an earlier version of this
patch.

MFC r334286:

Use correct mask.
Introduced in https://svnweb.freebsd.org/changeset/base/333603.
Thanks to Irene Ruengler for testing and reporting the issue.


git-svn-id: https://svn.freebsd.org/base/stable/11@334729 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f